Tests/reorganize: move denial of service tests to own file
